factor out utility functions
